Refactor types to own file
